+++ plone/outputfilters/testing.py
@@ -5,12 +5,12 @@ from plone.app.testing import applyProfile
 from plone.app.contenttypes.testing import PLONE_APP_CONTENTTYPES_FIXTURE
 
 from plone.outputfilters.filters.resolveuid_and_caption import IImageCaptioningEnabler  # noqa
-from zope.interface import implements
+from zope.interface import implementer
 import zope.component
 
 
+@implementer(IImageCaptioningEnabler)
 class DummyImageCaptioningEnabler(object):
-    implements(IImageCaptioningEnabler)
 
     available = True